+++ Goonies/Goonies/LockedDoor.cpp
@@ -48,7 +48,7 @@ void LockedDoor::render()
 void LockedDoor::interact()
 {
 	glm::vec2 lockPosition = glm::vec2(position.x, position.y + 2 * map->getTileSize()); //el candau esta 2 tile per sota
-	if (inContactWithPlayer(lockPosition, glm::vec2(8, 8))) {
+	if (inContactWithPlayer(lockPosition, glm::ivec2(8, 8))) {
 		if (player->useKey()) {
 			open();
 		}
